add Drawable shape superclass and refactor slightly
Rename Shape#outline() to Shape#points(). This allows for the soon to be
created FilledSquare class.

Should probably rethink the whole hierarchy at some point...

Add Drawable superclass to shape classes. This is almost certainly bad
as the name describes behaviour not identity, but fuck it. We can
refactor this after a proper discussion.